pyproject.toml
[tool.poetry]
name = "ya-gpt-bot"
version = "0.1.0"
description = "YandexGPT Telegram Bot"
authors = ["Aleksei Sokol <kanootoko@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.8"
click = "^8.1.7"
sqlalchemy = "^2.0.22"
pydantic = "^2.7"
requests = "^2.31.0"
loguru = "^0.7.2"
aiohttp = "^3.9"
asyncpg = "^0.29.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pylint = "^3.0.1"
pre-commit = "^3.5.0"
black = "^23.10.0"
isort = "^5.12.0"
alembic = "^1.12.0"
psycopg2 = "^2.9.9"
ipykernel = "^6.27.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"

[tool.poetry.scripts]
"ya-gpt-bot" = "ya_gpt_bot.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.pylint]
ignored-modules = ["alembic.context", "loguru"]
generated-members = ["sqlalchemy"]
extension-pkg-allow-list = []
max-line-length = 120
disable = ["duplicate-code"]

[tool.isort]
line_length = 120
profile = "black"

[tool.pytest.ini_options]
testpaths = ["tests"]
